Use new useStory composable and fix up routes.
