+++ app/pages/stories.vue
@@ -2,9 +2,21 @@
 import type { PageStoryblok } from '@@/types/storyblok'
 
 const route = useRoute()
-const story = await useStoryblokStory<PageStoryblok>('/stories')
+const story = await useStory<PageStoryblok>('/stories')
+const { title, description, image } = story.value.content.seo?.[0] ?? {}
+const imageOptions = { width: 1200, height: 630, format: 'jpg', smart: true, quality: 90 }
 
-useStoryblokSetup<PageStoryblok>(story)
+useSeoMeta({
+  title: title ?? story.value.name,
+  description,
+  ogTitle: title ?? story.value.name,
+  ogDescription: description,
+  ogImage: storyblokImage(image?.filename, imageOptions) || null,
+  ogType: 'website',
+  twitterTitle: title ?? story.value.name,
+  twitterCard: 'summary_large_image',
+  twitterImage: storyblokImage(image?.filename, imageOptions) || null,
+})
 </script>
 
 <template>
